+++ tests/index.spec.ts
@@ -1,8 +1,9 @@
 /// <reference path="../src/core.d.ts" />
-import { FormStateSpec } from "./spec/formstate.spec";
+import FormStateSpec from "./spec/formstate.spec";
 import ViewInternalSpec from "./spec/view-internal.spec";
 import ViewSpec from "./spec/view.spec";
-import { FormViewSpec } from "./spec/formview.spec";
+import SubviewSpec from "./spec/subview.spec";
+import FormViewSpec from "./spec/formview.spec";
 import UtilsSpec from "./spec/utils.spec";
 
 
@@ -10,4 +11,5 @@ UtilsSpec();
 FormStateSpec();
 ViewInternalSpec();
 ViewSpec();
+SubviewSpec();
 FormViewSpec();