+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## 0.1.3
 * Introducing React lifecycle methods componentWillMount, componentDidMount, shouldComponentUpdate, componentWillUpdate, componentDidUpdate.
+* Two-way view communication: parent view to child one and child to parent